chore: rename pkg to core
