+++ core/logging/std_logger.go
