+++ docs/src/charmil_factory.md
@@ -1,7 +1,7 @@
 ## Charmil Factory
 The Charmil Factory offers you one-stop access to all of the Charmil packages all at once. So you just need to initialize factory in your cobra command and start using logging, i18n, etc. 
 ```go
-import github.com/aerogear/charmil/pkg/factory
+import github.com/aerogear/charmil/core/factory
 ```
 
 ## How to use
